+++ src/app/admin/(configuration)/modules/types/modulesTypes.ts
@@ -1,7 +1,7 @@
 export interface ModuleCRUDType {
   id: string;
   title: string;
-  description?: string;
+  description: string;
   icon: string | null;
   order: number;
   url?: string;
@@ -12,4 +12,14 @@ export interface Module {
   id: string;
   title: string;
   order: number;
+}
+
+export interface CUModuleType {
+  title: string;
+  url?: string;
+  icon?: string;
+  order: number;
+  idModule?: string;
+  status?: string;
+  description: string;
 }
